% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-afni.R
\name{afni_hrf}
\alias{afni_hrf}
\title{construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.}
\usage{
afni_hrf(
  ...,
  basis = c("spmg1", "block", "dmblock", "tent", "csplin", "poly", "sin", "sine", "gam",
    "gamma", "spmg2", "spmg3", "wav"),
  onsets = NULL,
  durations = NULL,
  prefix = NULL,
  subset = NULL,
  nbasis = 1,
  contrasts = NULL,
  id = NULL,
  lag = 0,
  precision = 0.3,
  summate = TRUE,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{start}{the start of the window for sin/poly/csplin models}

\item{stop}{the stop time for sin/poly/csplin models}
}
\value{
an \code{afni_hrfspec} instance
}
\description{
construct an native AFNI hrf specification for '3dDeconvolve' with the 'stim_times' argument.
}
