% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf-afni.R
\name{afni_trialwise}
\alias{afni_trialwise}
\title{construct a native AFNI hrf specification for '3dDeconvolve' and individually modulated events using the 'stim_times_IM' argument.}
\usage{
afni_trialwise(
  label,
  basis = c("spmg1", "block", "dmblock", "gamma", "wav"),
  onsets = NULL,
  durations = 0,
  subset = NULL,
  id = NULL,
  start = 0,
  stop = 22,
  precision = 0.3,
  summate = TRUE
)
}
\arguments{
\item{label}{name of regressor}

\item{start}{start of hrf (for multiple basis hrfs)}

\item{stop}{end of hrf (for multiple basis hrfs)}
}
\value{
an \code{afni_trialwise_hrfspec} instance
}
\description{
construct a native AFNI hrf specification for '3dDeconvolve' and individually modulated events using the 'stim_times_IM' argument.
}
\examples{


tw <- afni_trialwise("trialwise", basis="gamma", onsets=seq(1,100,by=5))

}
