% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/sampling_frame.R
\name{blockids}
\alias{blockids}
\alias{blockids.sampling_frame}
\title{Get block identifiers}
\usage{
blockids(x, ...)

\method{blockids}{sampling_frame}(x, ...)
}
\arguments{
\item{x}{Object containing block structure}

\item{...}{Additional arguments passed to methods}
}
\value{
Integer vector of block ids
}
\description{
Generic accessor returning block indices for each sample or onset.
}
\examples{
# Get block identifiers from a sampling frame
sframe <- sampling_frame(blocklens = c(100, 120, 80), TR = 2)
blockids(sframe)
}
