% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frame.R
\name{blocklens.sampling_frame}
\alias{blocklens.sampling_frame}
\title{Get block lengths from a sampling frame}
\usage{
blocklens.sampling_frame(x, ...)
}
\value{
Numeric vector giving the number of scans in each block.
}
\description{
Get block lengths from a sampling frame
}
