% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-methods.R
\name{deriv.SPMG1_HRF}
\alias{deriv.SPMG1_HRF}
\title{Derivative method for SPMG1 HRF}
\usage{
\method{deriv}{SPMG1_HRF}(x, t, ...)
}
\arguments{
\item{x}{An SPMG1_HRF object}

\item{t}{Numeric vector of time points at which to evaluate the derivative}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector of derivative values
}
\description{
Uses the analytic derivative formula for the SPM canonical HRF.
}
