% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/reg-methods.R
\name{durations}
\alias{durations}
\alias{durations.Reg}
\title{Get durations of an object}
\usage{
durations(x, ...)

\method{durations}{Reg}(x, ...)
}
\arguments{
\item{x}{The object to get durations from}

\item{...}{Additional arguments passed to methods}
}
\value{
A numeric vector of durations
}
\description{
Get durations of an object
}
\examples{
# Create a regressor with event durations
reg <- regressor(onsets = c(1, 5, 10), hrf = HRF_SPMG1,
                 duration = c(2, 3, 1), span = 20)
durations(reg)
}
