% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrihrf-package.R
\docType{package}
\name{fmrihrf-package}
\alias{fmrihrf}
\alias{fmrihrf-package}
\title{fmrihrf: Hemodynamic Response Functions for fMRI Data Analysis}
\description{
Creates, manipulates, and evaluates hemodynamic response functions and event-related regressors for functional magnetic resonance imaging data analysis. Supports multiple basis sets including Canonical, Gamma, Gaussian, B-spline, and Fourier bases. Features decorators for time-shifting and blocking, and efficient convolution algorithms for regressor construction. Methods are based on standard fMRI analysis techniques as described in Jezzard et al. (2001, ISBN:9780192630711).
}
\details{
The fmrihrf package provides tools for creating, manipulating, and evaluating
Hemodynamic Response Functions (HRFs) and event-related regressors
for fMRI data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/fmrihrf/}
  \item Report bugs at \url{https://github.com/bbuchsbaum/fmrihrf/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
