% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hrf.R, R/reg-methods.R
\name{nbasis}
\alias{nbasis}
\alias{nbasis.HRF}
\alias{nbasis.Reg}
\title{Number of basis functions}
\usage{
nbasis(x, ...)

\method{nbasis}{HRF}(x, ...)

\method{nbasis}{Reg}(x, ...)
}
\arguments{
\item{x}{Object containing HRF or regressor information.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Integer scalar giving the number of basis functions.
}
\description{
Return the number of basis functions represented by an object.
}
\details{
This information is typically used when constructing penalty matrices
or understanding the complexity of an HRF model or regressor.
}
