% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf_decorators.R
\name{normalise_hrf}
\alias{normalise_hrf}
\title{Normalise Each Basis of an HRF to Unit Peak}
\usage{
normalise_hrf(hrf)
}
\arguments{
\item{hrf}{An object of class `HRF`.}
}
\value{
A unit-peak `HRF` object.
}
\description{
Back-compatible spelling and behavior for independently peak-normalizing
every basis column. New code can use
`normalize_hrf(hrf, "unit_peak_per_basis")` explicitly.
}
\examples{
gauss_unnorm <- as_hrf(function(t) 5 * dnorm(t, 6, 2), name = "unnorm_gauss")
gauss_norm <- normalise_hrf(gauss_unnorm)
max(gauss_norm(seq(0, 20, by = 0.1)))
}
\seealso{
Other HRF_decorator_functions: 
\code{\link{block_hrf}()},
\code{\link{lag_hrf}()},
\code{\link{normalize_hrf}()}
}
\concept{HRF_decorator_functions}
