% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-methods.R, R/sampling_frame.R
\name{print.Reg}
\alias{print.Reg}
\alias{print.sampling_frame}
\title{Print method for Reg objects}
\usage{
\method{print}{Reg}(x, ...)

\method{print}{sampling_frame}(x, ...)
}
\arguments{
\item{x}{A `Reg` object.}

\item{...}{Not used.}
}
\description{
Provides a concise summary of the regressor object using the cli package.
}
