% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-methods.R
\name{deriv.HRF}
\alias{deriv.HRF}
\title{Default derivative method for HRF objects}
\usage{
\method{deriv}{HRF}(x, t, ...)
}
\arguments{
\item{x}{An HRF object}

\item{t}{Numeric vector of time points at which to evaluate the derivative}

\item{...}{Additional arguments (currently unused)}
}
\value{
Numeric vector or matrix of derivative values
}
\description{
Uses numerical differentiation via numDeriv::grad when analytic derivatives
are not available for a specific HRF type.
}
