% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/reg-methods.R
\name{durations}
\alias{durations}
\alias{durations.Reg}
\title{Get durations of an object}
\usage{
durations(x, ...)

\method{durations}{Reg}(x, ...)
}
\arguments{
\item{x}{The object to get durations from}

\item{...}{Additional arguments passed to methods}
}
\value{
A numeric vector of durations
}
\description{
Get durations of an object
}
