% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hrf.R, R/reg-methods.R
\name{nbasis}
\alias{nbasis}
\alias{nbasis.HRF}
\alias{nbasis.Reg}
\title{Number of basis functions}
\usage{
nbasis(x, ...)

\method{nbasis}{HRF}(x, ...)

\method{nbasis}{Reg}(x, ...)
}
\description{

}
