% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrf.R
\name{plot.HRF}
\alias{plot.HRF}
\title{Plot an HRF Object}
\usage{
\method{plot}{HRF}(x, ...)
}
\arguments{
\item{x}{An HRF object}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot an HRF Object
}
